requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.5.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-celery==3.2.2
django-cors-headers==2.4.0
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-rest-swagger==2.2.0
django-simpleui==2020.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kombu==3.0.37
loguru==0.5.1
MarkupSafe==2.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==21.3
pydash==5.0.2
PyJWT==1.7.1
pyparsing==3.0.7
pytz==2022.1
PyYAML==5.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
